model_serving/reranker/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
loguru==0.7.3
vllm==0.9.2
python-dotenv==1.1.1